pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gaia-discovery-v3"
version = "0.1.0"
description = "gaia-discovery v3: Claude Code 驱动的 Gaia DSL 探索系统"
requires-python = ">=3.12"
dependencies = [
  "gaia-lang",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "httpx>=0.27",
  "typer[all]>=0.12",
  "pydantic>=2.0",
  "pydantic-settings>=2.0",
  "libcst>=1.4",
  "rich>=13.7",
  "mcp>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
  "ruff>=0.6.0",
]

[project.scripts]
gd = "gd.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/gd"]

[tool.hatch.build.targets.wheel.force-include]
"skills" = "gd/skills"
"templates" = "gd/templates"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = [
  "src",
  "/root/Gaia",
]
markers = [
  "llm: tests that require real LLM endpoint credentials",
  "lean: tests that require Lean toolchain",
  "claude_cli: tests that fork claude -p subprocess",
  "e2e: end-to-end smoke tests",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "W"]
ignore = ["E501"]
